+++ src/app/cfp/page.tsx
@@ -1,7 +1,7 @@
 import { CallToActionProps, Hero, ImageProps } from "@/sections/Hero/Hero";
 import { Header } from "@/components/Header/Header";
-import { cfp as data, footer } from "@/lib/data";
-import { getMetaData, getViewports } from "@/lib/metadata";
+import { cfp as data, footer , getMetaData, getViewports } from "@/lib";
+
 
 export const generateMetadata = () =>
   getMetaData({
